refector structure
